Bug: fix catalog app same version issue

+++ pkg/nirmata/resource_catalog_run_app.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	"math/rand"
+	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/nirmata/go-client/pkg/client"
@@ -70,35 +71,31 @@ func resourceRunApplicationCreate(d *schema.ResourceData, meta interface{}) erro
 	environments := d.Get("environments").([]interface{})
 	clusters := d.Get("clusters").([]interface{})
 
-	versionID, err := apiClient.QueryByName(client.ServiceCatalogs, "Version", version)
-	if err != nil {
-		fmt.Printf("Error - %v", err)
-		return err
-	}
-	catalogID, appErr := apiClient.QueryByName(client.ServiceCatalogs, "Catalogs", catalog)
-	if appErr != nil {
-		log.Printf("Error  - %v", appErr)
-		return err
+	catalogID, catErr := apiClient.QueryByName(client.ServiceCatalogs, "Catalogs", catalog)
+	if catErr != nil {
+		log.Printf("Error: catalog not found - %v", catErr)
+		return catErr
 	}
 	fields := []string{"name", "id"}
 	applicationList, appErr := apiClient.GetDescendants(catalogID, "Application", &client.GetOptions{Fields: fields})
 	if appErr != nil {
-		log.Printf("Error - %v", appErr)
-		return err
+		log.Printf("Error application not found - %v", appErr)
+		return appErr
 	}
 
 	var applicationID client.ID
 	var channelID string
+	var versionID string
 	for _, apps := range applicationList {
 		if application == apps["name"] {
 			applicationID = client.NewID(client.ServiceCatalogs, "Application", apps["id"].(string))
 		}
 	}
 
-	channelList, appErr := apiClient.GetDescendants(applicationID, "Channel", &client.GetOptions{Fields: fields})
-	if appErr != nil {
-		log.Printf("Error - %v", appErr)
-		return err
+	channelList, channelErr := apiClient.GetDescendants(applicationID, "Channel", &client.GetOptions{Fields: fields})
+	if channelErr != nil {
+		log.Printf("Error channel not found - %v", channelErr)
+		return channelErr
 	}
 
 	for _, c := range channelList {
@@ -111,11 +108,28 @@ func resourceRunApplicationCreate(d *schema.ResourceData, meta interface{}) erro
 		"service":    "Catalog",
 		"modelIndex": "Channel",
 	}
+
+	versionList, versionErr := apiClient.GetDescendants(applicationID, "Version", &client.GetOptions{Fields: fields})
+	if versionErr != nil {
+		log.Printf("Error version not found - %v", versionErr)
+		return versionErr
+	}
+
+	for _, c := range versionList {
+		if version == c["name"] {
+			versionID = c["id"].(string)
+		}
+	}
+
 	var catalogEnvArr = make([]interface{}, 0)
 	var catalogClusterArr = make([]interface{}, 0)
 	if len(environments) != 0 {
 		for _, env := range environments {
-			envId, _ := apiClient.QueryByName(client.ServiceEnvironments, "Environment", env.(string))
+			envId, envErr := apiClient.QueryByName(client.ServiceEnvironments, "Environment", env.(string))
+			if envErr != nil {
+				log.Printf("Error environment not found - %v", envErr)
+				return envErr
+			}
 			obj := map[string]interface{}{
 				"id":         envId.UUID(),
 				"service":    "Environments",
@@ -126,8 +140,12 @@ func resourceRunApplicationCreate(d *schema.ResourceData, meta interface{}) erro
 	}
 
 	if len(clusters) != 0 {
-		for _, env := range environments {
-			cluId, _ := apiClient.QueryByName(client.ServiceClusters, "KubernetesCluster", env.(string))
+		for _, cluster := range clusters {
+			cluId, cErr := apiClient.QueryByName(client.ServiceClusters, "KubernetesCluster", cluster.(string))
+			if cErr != nil {
+				log.Printf("Error cluster not found - %v", cErr)
+				return cErr
+			}
 			obj := map[string]interface{}{
 				"id":         cluId.UUID(),
 				"service":    "Cluster",
@@ -140,7 +158,7 @@ func resourceRunApplicationCreate(d *schema.ResourceData, meta interface{}) erro
 	txnData := map[string]interface{}{
 		"runName":    name,
 		"name":       name + "-" + fmt.Sprint(rand.Int()),
-		"parent":     versionID.UUID(),
+		"parent":     versionID,
 		"envIds":     catalogEnvArr,
 		"clusterIds": catalogClusterArr,
 		"channel":    channelMap,
@@ -155,6 +173,23 @@ func resourceRunApplicationCreate(d *schema.ResourceData, meta interface{}) erro
 
 	changeID := rolloutId["id"].(string)
 	d.SetId(changeID)
+
+	rolloutID := client.NewID(client.ServiceCatalogs, "Rollout", changeID)
+	state, waitErr := waitForRolloutState(apiClient, d.Timeout(schema.TimeoutCreate), rolloutID)
+	if waitErr != nil {
+		log.Printf("[ERROR] - failed to check rollout status. Error - %v", waitErr)
+		return waitErr
+	}
+
+	if strings.EqualFold("failed", state) {
+		status, err := getRolloutStatus(apiClient, rolloutID)
+		if err != nil {
+			log.Printf("[ERROR] - failed to retrieve rollout failure details: %v", err)
+			return fmt.Errorf("rollout creation failed")
+		}
+		return fmt.Errorf("rollout creation failed: %s", status)
+	}
+
 	log.Printf("[INFO] - created  rollout %s %s", name, changeID)
 
 	return nil
@@ -172,12 +207,35 @@ func resourceRunApplicationUpdate(d *schema.ResourceData, meta interface{}) erro
 func resourceRunApplicationDelete(d *schema.ResourceData, meta interface{}) error {
 	apiClient := meta.(client.Client)
 	name := d.Get("name").(string)
-	id := client.NewID(client.ServiceCatalogs, "Rollout", d.Id())
+	environments := d.Get("environments").([]interface{})
+	application := d.Get("application").(string)
+	fields := []string{"name", "id"}
+	if len(environments) != 0 {
+		for _, env := range environments {
+			envId, envErr := apiClient.QueryByName(client.ServiceEnvironments, "Environment", env.(string))
+			if envErr != nil {
+				log.Printf("Error environment not found - %v", envErr)
+				return envErr
+			}
+			applicationList, appErr := apiClient.GetDescendants(envId, "Application", &client.GetOptions{Fields: fields})
+			if appErr != nil {
 
-	if err := apiClient.Delete(id, nil); err != nil {
-		return err
+				log.Printf("Error applications not found - %v", appErr)
+				return appErr
+			}
+			for _, apps := range applicationList {
+				if application == apps["name"] {
+					applicationID := client.NewID(client.ServiceEnvironments, "Application", apps["id"].(string))
+					id := client.NewID(client.ServiceEnvironments, "Application", apps["id"].(string))
+					if err := apiClient.Delete(id, nil); err != nil {
+						return err
+					}
+					waitForDeletedState(apiClient, d.Timeout(schema.TimeoutCreate), applicationID)
+				}
+			}
+		}
 	}
 
-	log.Printf("[INFO] - deleted rollout %s %s", name, id.UUID())
+	log.Printf("[INFO] - deleted rollout %s", name)
 	return nil
 }
